Summer sweetness 2023

+++ _plugins/tailwind_css.rb
@@ -1,11 +1,14 @@
+# frozen_string_literal: true
+
 module TailwindCss
+  # This is the main engine that processes the css.
   class Engine
     def initialize(bundle, destination)
       @bundle = bundle
       @input = bundle['input']
       @destination = destination
       @output = ".jekyll-cache/#{input}"
-      @minified = !!bundle['minified']
+      @minified = bundle.fetch('minified', false)
       @config = bundle['config'] ? "--config #{bundle['config']}" : nil
     end
 
@@ -13,14 +16,8 @@ module TailwindCss
       contents = File.read(input).gsub(/^---$/, '')
       tempfile.write contents
       tempfile.rewind
-
-      `tailwindcss --input #{tempfile.path} --output #{output} #{config}`
-      Jekyll.logger.info "Tailwind:", "Processed #{input}"
-
-      if minified
-        `parcel-css --minify --nesting #{output} -o #{output}`
-        Jekyll.logger.info "Tailwind:", "Minified output"
-      end
+      run_tailwind
+      run_parcel if minified
     ensure
       tempfile.close
       tempfile.unlink
@@ -30,18 +27,17 @@ module TailwindCss
       final_path = "#{destination}/#{input}"
       final_dir = File.dirname final_path
 
-      if File.exist? output
-        FileUtils.mkdir_p final_dir
-        FileUtils.cp output, final_path
-        Jekyll.logger.info "Tailwind:", "Copied to #{final_path}"
-      end
+      return unless File.exist? output
+
+      FileUtils.mkdir_p final_dir
+      FileUtils.cp output, final_path
+      Jekyll.logger.info 'Tailwind:', "Copied to #{final_path}"
     end
 
     def cleanup
-      FileUtils.rm output
-      Jekyll.logger.info "Tailwind:", "Cleaned out cached file"
+      FileUtils.rm output if File.exist? output
+      Jekyll.logger.info 'Tailwind:', 'Cleaned out cached file'
     end
-
 
     protected
 
@@ -49,8 +45,18 @@ module TailwindCss
 
     private
 
+    def run_tailwind
+      `tailwindcss --input #{tempfile.path} --output #{output} #{config}`
+      Jekyll.logger.info 'Tailwind:', "Processed #{input}"
+    end
+
+    def run_parcel
+      `parcel-css --minify --nesting #{output} -o #{output}`
+      Jekyll.logger.info 'Tailwind:', 'Minified output'
+    end
+
     def tempfile
-      @_tempfile ||= Tempfile.new(input)
+      @tempfile ||= Tempfile.new(input)
     end
   end
 end
@@ -62,11 +68,11 @@ Jekyll::Hooks.register :pages, :post_init do |page|
   css_bundles = page.site.config['tailwindcss']
 
   css_bundles.each do |bundle|
-    if /\.css$/ =~ page.url && page.url.include?(bundle['input'])
-      engine = TailwindCss::Engine.new(bundle, page.site.config['destination'])
-      engine.process
-      engine.copy
-    end
+    next unless /\.css$/ =~ page.url && page.url.include?(bundle['input'])
+
+    engine = TailwindCss::Engine.new(bundle, page.site.config['destination'])
+    engine.process
+    engine.copy
   end
 end
 
@@ -75,11 +81,10 @@ Jekyll::Hooks.register :pages, :post_write do |page|
   css_bundles = page.site.config['tailwindcss']
 
   css_bundles.each do |bundle|
-    if /\.css$/ =~ page.url && page.url.include?(bundle['input'])
-      engine = TailwindCss::Engine.new(bundle, page.site.config['destination'])
-      engine.copy
-      engine.cleanup
-    end
+    next unless /\.css$/ =~ page.url && page.url.include?(bundle['input'])
+
+    engine = TailwindCss::Engine.new(bundle, page.site.config['destination'])
+    engine.copy
+    engine.cleanup
   end
 end
-
